Migrate community settings recipient page to use dynamic routing #362
