+++ routes/logged-in/sidebar/community-settings-recipient/index.js
@@ -1,0 +1,11 @@
+// polyfill webpack require.ensure
+if (typeof require.ensure !== 'function') require.ensure = (d, c) => c(require)
+
+export default store => ({
+  path: 'community/recipient',
+  getComponent (nextState, callback) {
+    require.ensure([], function (require) {
+      callback(null, require('./page.connected').default)
+    })
+  }
+})
